Managing more State - 2
The createStore() function we built can only handle a single reducer function. We can't call createStore() passing it two reducer functions. We solved this problem in this commit by having a root reducer

We have two reducer functions and each of the function is expected to receive their slice of the state tree whenever an action is dispatched. Since the createStore() method can take only one reducer, we will have to create almost like a root reducer function(A reducer that acts as the main reducer) which is responsible for calling the correct reducer to get each piece of state to build the final state of the app whenever specific actions are dispatched.

For this COMMIT:

We also created the goals reducer which handles:
- an ADD_GOAL action
- a REMOVE_GOAL action
So our application can now manage the state of our todos and goals, and it can do all of this, predictably!