Redux Middleware
The goal of the checker() method is to check for if the action payload data for the ADD_GOAL or ADD_TODO and check if the name property in the payload contains the word bitcoin

We could have our own middleware the way we created our own State Management library but it will get messy and unmaintenable when we start to have more than one middleware functions to call before we actually get to finally call the reducer method

What's great about middleware is that once it receives the action, it can carry out a number of operations, including:

- producing a side effect (e.g., logging information about the store)
- processing the action itself (e.g., making an asynchronous HTTP request)
- redirecting the action (e.g., to another piece of middleware)
- dispatching supplementary actions

…or even some combination of the above! Middleware can do any of these before passing the action along to the reducer.

Here's applyMiddleware()'s signature:
applyMiddleware(...middlewares)
Note the spread operator on the middlewares parameter. This means that we can pass in as many different middleware as we want!

Redux middleware leverages a concept called higher-order functions. A higher-order function is a function that either:
- accepts a function as an argument
- returns a function
Higher-order functions are a powerful programming technique that allow functions to be significantly more dynamic. You've actually already written a higher-order function in this course. The createRemoveButton() function is a higher-order function because the onClick parameter is expected to be a function (because onClick is set up as an event listener callback function.