Handling Initial Data
We needed a way to tell our Redux store about this data from the API. One way is looping through all of our goals and actions but we will be dispatching a lot of actions

How do we want to change the state of our store based on the RECEIVE_DATA action?

We ended up creating a single action type that will affects multiple parts of our store because when RECEIVE_DATA is dispatched we update the todos and goals slices of our store respectively.

So now, when we refresh our page, we will have the data initial data that comes from the database. The is a delay of 2 secs before the data loads because we set that on purpose in our fake DB