Optimistically Deleting Items
We dont want to wait until for a confirmation that the item is deleted from the database. We want to do that immediately.
We first call the dispatch before we make the API request.
If there is an error during the API request, we catch the error and add the item back to the list

You are still updating your database but the user is getting instant feedback when they click on the delete item