Custom Thunk
We mixed all our data fetching logic with our UI logic.
Ideally you will want your UI logic to just focus on how the UI looks and should not be responsible for fetching data. So you have your UI logic and data fetching logic separate

To solve this, instead of calling our API and then passing that data to our action creators like we doing, we would move the data fetching logic from our component to our action creators.

By calling our API in the action creator, we make the action creator responsible for fetching the data it needs to create the actual action.

- Remember that redux middleware sits between the dispatch and the reducer. An asynchronous request can be done in this middleware
- So what we did was instead of calling the dispatch directly and passing it in the action creator, we created a function that returns a dispatch method.
This method that returns a dispatch method does the api operation in an optimistic way(meaning, it will return the first dispatch action which updates the state, the it optionally will return a rollback dispatch action for the second time if the api fails)
- Now, we had to create a custom middleware(thunk) that will handle this type of situation for our handler. Basically, custom middleware checks if the action passed in is already a function(a dispatch function to be specific), the it will invoke the function(now we run the asynchronous API operation inside Redux middleware), else, it will dispatch as it will normally would(no API call, just normal dispatch)
We ended up using official Redux think code from the Redux Think package but the implementation is thesame to our custom think code

- NOTE that this function invoked inside the middleware that makes the API call, will still return invoke a dispatch function.