Build custom abstraction for Provider, Context, and connect() function
Remember the goal is to make React and Redux work more effectively together to make our code more maintainable increase predictability of ur APP.
To achieve this we will create various abstractions.
These abstractions helps us understand the code library downloaded from NPM that we will ideally use to achieve thesame goal. So think of these abstractions as a library code
So just as we built an abstraction over Context.Provider using Provider Component, we will also build an abstraction over Context.Consumer by building a connect function

In this commit, instead of putting our store on Context and grabbing that information from context ourselves(like which we did in the last commit), we created a little abstract helper library our Provider component and Connect higher order component. And now whenever want to grab something on context or off context, instead of doing that ourselves, we can now use our abstractions

Two abstractions that will eventually be a library download from NPM
- Provider component that is responsible for sticking our Redux store on context so that this way any child component that needs it can just grab it off context.
- The connect function: This connect function still needs to render any component and pass that component any data it needs from the store. More explanation of the connect function below
- These two abstractions helps us make react Redux work together effectively!!

The connect function
- It looks like this connect()()
- This connect function still needs to perform what our ConnectedComponent does because it is an abstract of it. So this connect functions still needs to render any component and pass that component any data it needs from the store.
- We used thesame technique that we used to implement handler for middleware functions where we have a function connect() return a function () that will be invoked by the second set of brackets
- In the first set of brackets we will pass in any data that the component needs, then we pass in the component that will actually be rendered in the second brackets
- Things to keep in mind when implementing this connect function is that we need to use store.subscribe somewhere in our code so that we are aware whenever the store changes; secondly figure out what state that the component that we will be rendering needs(we can get the store from the first bracket) as well as what we will render(the component passed in the second bracket) and lastly connect to context so that we can grab the store from the context

When invoking the connect function
In this bracket we pass in a function. This function will be passed the stroe as its argument and any property we return in this function will be passed as prop to the component we want to render in the second bracket

FYI: in your code, you see state as argument for the connect function but we will pass the state of the store in there